Move WithMeta class into separate file
