+++ optionsfactory/__init__.py
@@ -1,9 +1,11 @@
 from .optionsfactory import OptionsFactory
+from .withmeta import WithMeta
 from ._version import get_versions
 
 __version__ = get_versions()["version"]
 
 __all__ = [
     "OptionsFactory",
+    "WithMeta",
     "__version__",
 ]
